+++ app/templates/components/dynamic-questions.hbs
@@ -1,0 +1,11 @@
+<div id='input' class="test"> 
+	
+{{#each rangeOfQuestions as |index|}}
+
+	{{ survey-question }}
+
+{{/each}}
+     
+<button {{action 'add'}}>insert question </button>
+
+</div>
