+++ app/templates/components/survey-question.hbs
@@ -1,18 +1,30 @@
-<div id='input' class="test"> 
-	
-{{#each rangeOfQuestions as |index|}}
-
-	<p><strong> Question {{index}} </strong></p>
-	{{#em-form model=none}}
-	    {{em-input label="Title" property="title" cid="FID"}} 
-  		{{em-text property="help" label="Help" placeholder="Optional help..." rows=2}}
-	    {{em-checkbox property="required" label="Required?"}}
-	{{/em-form}}
-
-{{/each}}
-     
-<button {{action 'add'}}>insert question </button>
-
+<div class="panel panel-default">
+  <div class="panel-heading">
+    <h3 class="panel-title">Question {{ index }}</h3>
+  </div>
+  <div class="panel-body">
+    <form>
+	  <div class="form-group">
+	    <label for="inputTitle">Title</label>
+	    <input type="text" class="form-control" id="inputTitle"
+	      placeholder="Title">
+	  </div>
+	  <div class="form-group">
+	    <label for="inputQuestion">Question</label>
+	    <textarea class="form-control" id="inputQuestion"
+	      placeholder="Question"></textarea>
+	  </div>
+	  <div class="form-group">
+	    <label for="inputHelp">Help</label>
+	    <textarea class="form-control" id="inputHelp"
+	      placeholder="Optional help..."></textarea>
+	  </div>
+	  <div class="form-group">
+	    <label for="required">Required</label>
+	    <input type="checkbox" class="form-control" id="required"
+	      placeholder=False>
+	  </div>
+	  <button type="submit" class="btn btn-default">Submit</button>
+	</form>
+  </div>
 </div>
-
-
