+++ app/templates/dynamic.hbs
@@ -1,1 +1,1 @@
-{{ survey-question }}
+{{ dynamic-questions }}
