+++ app/templates/question.hbs
@@ -1,5 +1,1 @@
-{{#em-form model=model formLayout=horizontal}}
-  {{em-input label="Title" property="title" cid="FID"}} 
-  {{em-text property="help" label="Help" placeholder="Optional help..." rows=2}}
-  {{em-checkbox property="required" label="Required?"}}
-{{/em-form}}
+{{ survey-question }}
