+++ tests/integration/components/dynamic-questions-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('dynamic-questions', 'Integration | Component | dynamic questions', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });"
+
+  this.render(hbs`{{dynamic-questions}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:"
+  this.render(hbs`
+    {{#dynamic-questions}}
+      template block text
+    {{/dynamic-questions}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
